Remove / Replace image
Hello Batami. 

I kindly ask for the image shown on this article to be removed / replaced due to the fact that the client being shown is the Office Built-in Solution. According to https://docs.microsoft.com/en-us/azure/information-protection/rms-client/use-client#compare-the-labeling-solutions-for-windows-computers we can see that this option is only available for the UL client in preview. The Built-In solution does not support it accorrding to the mentioned article. 

Thanks in advance:)